Created SQL script
In SQL query, if the differences need to be found between two specific dated like Order Date and Shipment date, DATEDIFF function could used for that. The issue was both of column which show dates are not DATETIME format, so that they needed be converted as DATETIME format.  It fixed and solved the differences of between days