Created SQL Script for Practice
Views are virtual tables based on the result set of an SQL statement.
Views can be used as security mechanisms by letting users access data through the view, without granting users permissions to directly access the underlying tables of the query. Views can be used to provide a backward compatible interface to emulate a table that used to exist but whose schema has changed.